// src/app/@sunflower-module/sunflower-ui/model/personalDetailsDTO.ts
/**
 * Blue Frameworks
 * Spring Framework 5
 *
 * OpenAPI spec version: 1.0
 * Contact: luke.petzer@younglings.africa
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface PersonalDetailsDTO {
    address1?: string;
    address2?: string;
    address3?: string;
    birthDate?: string;
    countryId?: number;
    ethnicGroupId?: string;
    firstContactEmail?: string;
    firstContactMobile?: string;
    firstContactName?: string;
    firstContactRelationship?: number;
    firstName?: string;
    gender?: string;
    homePhone?: string;
    idNumber?: string;
    idType?: number;
    postalCode?: string;
    provinceId?: number;
    secondContactEmail?: string;
    secondContactMobile?: string;
    secondContactName?: string;
    secondContactRelationship?: number;
    surname?: string;
    titleId?: number;
    workPhone?: string;
}